api,as: Move webhook templates gRPC endpoints to the webhook registry
