+++ pkg/ttnpb/applicationserver_web.pb.validate.go
@@ -759,206 +759,6 @@ var _ interface {
 	ErrorName() string
 } = ApplicationWebhookTemplatesValidationError{}
 
-// ValidateFields checks the field values on
-// GetApplicationWebhookTemplateRequest with the rules defined in the proto
-// definition for this message. If any rules are violated, an error is returned.
-func (m *GetApplicationWebhookTemplateRequest) ValidateFields(paths ...string) error {
-	if m == nil {
-		return nil
-	}
-
-	if len(paths) == 0 {
-		paths = GetApplicationWebhookTemplateRequestFieldPathsNested
-	}
-
-	for name, subs := range _processPaths(append(paths[:0:0], paths...)) {
-		_ = subs
-		switch name {
-		case "ids":
-
-			if v, ok := interface{}(&m.ApplicationWebhookTemplateIdentifiers).(interface{ ValidateFields(...string) error }); ok {
-				if err := v.ValidateFields(subs...); err != nil {
-					return GetApplicationWebhookTemplateRequestValidationError{
-						field:  "ids",
-						reason: "embedded message failed validation",
-						cause:  err,
-					}
-				}
-			}
-
-		case "field_mask":
-
-			if v, ok := interface{}(&m.FieldMask).(interface{ ValidateFields(...string) error }); ok {
-				if err := v.ValidateFields(subs...); err != nil {
-					return GetApplicationWebhookTemplateRequestValidationError{
-						field:  "field_mask",
-						reason: "embedded message failed validation",
-						cause:  err,
-					}
-				}
-			}
-
-		default:
-			return GetApplicationWebhookTemplateRequestValidationError{
-				field:  name,
-				reason: "invalid field path",
-			}
-		}
-	}
-	return nil
-}
-
-// GetApplicationWebhookTemplateRequestValidationError is the validation error
-// returned by GetApplicationWebhookTemplateRequest.ValidateFields if the
-// designated constraints aren't met.
-type GetApplicationWebhookTemplateRequestValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e GetApplicationWebhookTemplateRequestValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e GetApplicationWebhookTemplateRequestValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e GetApplicationWebhookTemplateRequestValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e GetApplicationWebhookTemplateRequestValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e GetApplicationWebhookTemplateRequestValidationError) ErrorName() string {
-	return "GetApplicationWebhookTemplateRequestValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e GetApplicationWebhookTemplateRequestValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sGetApplicationWebhookTemplateRequest.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = GetApplicationWebhookTemplateRequestValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = GetApplicationWebhookTemplateRequestValidationError{}
-
-// ValidateFields checks the field values on
-// ListApplicationWebhookTemplatesRequest with the rules defined in the proto
-// definition for this message. If any rules are violated, an error is returned.
-func (m *ListApplicationWebhookTemplatesRequest) ValidateFields(paths ...string) error {
-	if m == nil {
-		return nil
-	}
-
-	if len(paths) == 0 {
-		paths = ListApplicationWebhookTemplatesRequestFieldPathsNested
-	}
-
-	for name, subs := range _processPaths(append(paths[:0:0], paths...)) {
-		_ = subs
-		switch name {
-		case "field_mask":
-
-			if v, ok := interface{}(&m.FieldMask).(interface{ ValidateFields(...string) error }); ok {
-				if err := v.ValidateFields(subs...); err != nil {
-					return ListApplicationWebhookTemplatesRequestValidationError{
-						field:  "field_mask",
-						reason: "embedded message failed validation",
-						cause:  err,
-					}
-				}
-			}
-
-		default:
-			return ListApplicationWebhookTemplatesRequestValidationError{
-				field:  name,
-				reason: "invalid field path",
-			}
-		}
-	}
-	return nil
-}
-
-// ListApplicationWebhookTemplatesRequestValidationError is the validation
-// error returned by ListApplicationWebhookTemplatesRequest.ValidateFields if
-// the designated constraints aren't met.
-type ListApplicationWebhookTemplatesRequestValidationError struct {
-	field  string
-	reason string
-	cause  error
-	key    bool
-}
-
-// Field function returns field value.
-func (e ListApplicationWebhookTemplatesRequestValidationError) Field() string { return e.field }
-
-// Reason function returns reason value.
-func (e ListApplicationWebhookTemplatesRequestValidationError) Reason() string { return e.reason }
-
-// Cause function returns cause value.
-func (e ListApplicationWebhookTemplatesRequestValidationError) Cause() error { return e.cause }
-
-// Key function returns key value.
-func (e ListApplicationWebhookTemplatesRequestValidationError) Key() bool { return e.key }
-
-// ErrorName returns error name.
-func (e ListApplicationWebhookTemplatesRequestValidationError) ErrorName() string {
-	return "ListApplicationWebhookTemplatesRequestValidationError"
-}
-
-// Error satisfies the builtin error interface
-func (e ListApplicationWebhookTemplatesRequestValidationError) Error() string {
-	cause := ""
-	if e.cause != nil {
-		cause = fmt.Sprintf(" | caused by: %v", e.cause)
-	}
-
-	key := ""
-	if e.key {
-		key = "key for "
-	}
-
-	return fmt.Sprintf(
-		"invalid %sListApplicationWebhookTemplatesRequest.%s: %s%s",
-		key,
-		e.field,
-		e.reason,
-		cause)
-}
-
-var _ error = ListApplicationWebhookTemplatesRequestValidationError{}
-
-var _ interface {
-	Field() string
-	Reason() string
-	Key() bool
-	Cause() error
-	ErrorName() string
-} = ListApplicationWebhookTemplatesRequestValidationError{}
-
 // ValidateFields checks the field values on ApplicationWebhook with the rules
 // defined in the proto definition for this message. If any rules are
 // violated, an error is returned.
@@ -1711,6 +1511,206 @@ var _ interface {
 	Cause() error
 	ErrorName() string
 } = SetApplicationWebhookRequestValidationError{}
+
+// ValidateFields checks the field values on
+// GetApplicationWebhookTemplateRequest with the rules defined in the proto
+// definition for this message. If any rules are violated, an error is returned.
+func (m *GetApplicationWebhookTemplateRequest) ValidateFields(paths ...string) error {
+	if m == nil {
+		return nil
+	}
+
+	if len(paths) == 0 {
+		paths = GetApplicationWebhookTemplateRequestFieldPathsNested
+	}
+
+	for name, subs := range _processPaths(append(paths[:0:0], paths...)) {
+		_ = subs
+		switch name {
+		case "ids":
+
+			if v, ok := interface{}(&m.ApplicationWebhookTemplateIdentifiers).(interface{ ValidateFields(...string) error }); ok {
+				if err := v.ValidateFields(subs...); err != nil {
+					return GetApplicationWebhookTemplateRequestValidationError{
+						field:  "ids",
+						reason: "embedded message failed validation",
+						cause:  err,
+					}
+				}
+			}
+
+		case "field_mask":
+
+			if v, ok := interface{}(&m.FieldMask).(interface{ ValidateFields(...string) error }); ok {
+				if err := v.ValidateFields(subs...); err != nil {
+					return GetApplicationWebhookTemplateRequestValidationError{
+						field:  "field_mask",
+						reason: "embedded message failed validation",
+						cause:  err,
+					}
+				}
+			}
+
+		default:
+			return GetApplicationWebhookTemplateRequestValidationError{
+				field:  name,
+				reason: "invalid field path",
+			}
+		}
+	}
+	return nil
+}
+
+// GetApplicationWebhookTemplateRequestValidationError is the validation error
+// returned by GetApplicationWebhookTemplateRequest.ValidateFields if the
+// designated constraints aren't met.
+type GetApplicationWebhookTemplateRequestValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e GetApplicationWebhookTemplateRequestValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e GetApplicationWebhookTemplateRequestValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e GetApplicationWebhookTemplateRequestValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e GetApplicationWebhookTemplateRequestValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e GetApplicationWebhookTemplateRequestValidationError) ErrorName() string {
+	return "GetApplicationWebhookTemplateRequestValidationError"
+}
+
+// Error satisfies the builtin error interface
+func (e GetApplicationWebhookTemplateRequestValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sGetApplicationWebhookTemplateRequest.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = GetApplicationWebhookTemplateRequestValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = GetApplicationWebhookTemplateRequestValidationError{}
+
+// ValidateFields checks the field values on
+// ListApplicationWebhookTemplatesRequest with the rules defined in the proto
+// definition for this message. If any rules are violated, an error is returned.
+func (m *ListApplicationWebhookTemplatesRequest) ValidateFields(paths ...string) error {
+	if m == nil {
+		return nil
+	}
+
+	if len(paths) == 0 {
+		paths = ListApplicationWebhookTemplatesRequestFieldPathsNested
+	}
+
+	for name, subs := range _processPaths(append(paths[:0:0], paths...)) {
+		_ = subs
+		switch name {
+		case "field_mask":
+
+			if v, ok := interface{}(&m.FieldMask).(interface{ ValidateFields(...string) error }); ok {
+				if err := v.ValidateFields(subs...); err != nil {
+					return ListApplicationWebhookTemplatesRequestValidationError{
+						field:  "field_mask",
+						reason: "embedded message failed validation",
+						cause:  err,
+					}
+				}
+			}
+
+		default:
+			return ListApplicationWebhookTemplatesRequestValidationError{
+				field:  name,
+				reason: "invalid field path",
+			}
+		}
+	}
+	return nil
+}
+
+// ListApplicationWebhookTemplatesRequestValidationError is the validation
+// error returned by ListApplicationWebhookTemplatesRequest.ValidateFields if
+// the designated constraints aren't met.
+type ListApplicationWebhookTemplatesRequestValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e ListApplicationWebhookTemplatesRequestValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e ListApplicationWebhookTemplatesRequestValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e ListApplicationWebhookTemplatesRequestValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e ListApplicationWebhookTemplatesRequestValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e ListApplicationWebhookTemplatesRequestValidationError) ErrorName() string {
+	return "ListApplicationWebhookTemplatesRequestValidationError"
+}
+
+// Error satisfies the builtin error interface
+func (e ListApplicationWebhookTemplatesRequestValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sListApplicationWebhookTemplatesRequest.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = ListApplicationWebhookTemplatesRequestValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = ListApplicationWebhookTemplatesRequestValidationError{}
 
 // ValidateFields checks the field values on ApplicationWebhookTemplate_Message
 // with the rules defined in the proto definition for this message. If any
