+++ pkg/ttnpb/field_mask_validation.go
@@ -65,14 +65,14 @@ var AllowedFieldMaskPathsForRPC = map[string][]string{
 	"/ttn.lorawan.v3.ApplicationRegistry/Update":              ApplicationFieldPathsNested,
 	"/ttn.lorawan.v3.EntityRegistrySearch/SearchApplications": ApplicationFieldPathsNested,
 
+	// Application Webhook Templates:
+	"/ttn.lorawan.v3.ApplicationWebhookRegistry/GetTemplate":   ApplicationWebhookTemplateFieldPathsNested,
+	"/ttn.lorawan.v3.ApplicationWebhookRegistry/ListTemplates": ApplicationWebhookTemplateFieldPathsNested,
+
 	// Application Webhooks:
 	"/ttn.lorawan.v3.ApplicationWebhookRegistry/Get":  ApplicationWebhookFieldPathsNested,
 	"/ttn.lorawan.v3.ApplicationWebhookRegistry/List": ApplicationWebhookFieldPathsNested,
 	"/ttn.lorawan.v3.ApplicationWebhookRegistry/Set":  ApplicationWebhookFieldPathsNested,
-
-	// Application Webhook Templates:
-	"/ttn.lorawan.v3.ApplicationWebhookTemplateRegistry/Get":  ApplicationWebhookTemplateFieldPathsNested,
-	"/ttn.lorawan.v3.ApplicationWebhookTemplateRegistry/List": ApplicationWebhookTemplateFieldPathsNested,
 
 	// Application PubSubs:
 	"/ttn.lorawan.v3.ApplicationPubSubRegistry/Get":  ApplicationPubSubFieldPathsNested,
